fix for CsvReader overflow on long lines

+++ DataProviders/Source/CsvReader.cpp
@@ -76,7 +76,7 @@ void DataProviders::CsvReader::setDelimiter(char delimiter) {
 
 std::vector<std::string> DataProviders::CsvReader::tokenize(std::string &line, int vectorReserve) {
   char buffer[BUFFER_SIZE];
-  uint8_t bufferIter = 0;
+  uint64_t bufferIter = 0;
   TokeniserStates state = Read; //0 - normal read, 1 - read inside "", 2 - found " inside ""
 
   std::vector<std::string> result;
